questionnaires_compiler/test/requirements-test.txt
pytest~=7.3
coverage~=7.3
pytest-django~=4.5
pytest-env~=1.1
pytest-xdist~=3.4
pytest-cov~=4.1.0
